Pull a specific feature set available on CaDrA Portal and determine whether to include gene expression set that is associated with that feature set

Pull a specific feature set available on CaDrA Portal and determine whether to include gene expression set that is associated with that feature set

pull_datasets(feature_set, include_gene_expression = TRUE)

Arguments

feature_set

a list of feature sets to look up and download the appropriate data that are associated with it.

include_gene_expression

a Boolean value to determine whether to include gene expression sets that associated with the given feature sets. Default is TRUE.

Value

a list of feature set and gene expression set by setting include_gene_expression = TRUE

Examples

# Retrieve a list of feature sets available from CaDrA Portal
fs_list <- CaDrA.shiny::get_feature_set(order_by = "asc")

# Retrieve the first feature set from fs_list and include its 
# associated gene expression set
datasets <- pull_datasets(
  feature_set = fs_list$feature_set_name[1],
  include_gene_expression = TRUE
)
#> Warning: cannot open compressed file '/var/folders/7b/hb_2pqcn5l79y3srmcdnj7bw0000gn/T//RtmpMXHp2D/download-fs-2023-11-27/CCLE SCNAs and Mutations/feature_set/CCLE_MUT_SCNA.rda', probable reason 'No such file or directory'
#> Error in gzfile(file, "rb"): cannot open the connection